refactor: extract CLI command handlers
